feat(core): text_time custom alphabets, wide encoding, and non-Unix epochs (ULID, KSUID)
The initial text_time build (previous commit) covered cuid v1 and MongoDB ObjectId --
both a plain contiguous 0-9a-z (or 0-9a-f) alphabet and a timestamp that IS the whole
decoded field. Two more common formats need genuinely more:

- ULID uses Crockford's base32 (0123456789ABCDEFGHJKMNPQRSTVWXYZ -- deliberately skips
  I/L/O/U), not the default alphabet. New p_tt_alphabet parameter: an explicit
  digit-to-character string overriding the default convention, with the default's radix
  cap (36) only applying when no alphabet is given -- a supplied alphabet's own length is
  the radix ceiling instead.
- KSUID base62-encodes its ENTIRE 160-bit payload (32-bit timestamp + 128 bits of random)
  as one number against a non-Unix epoch (2014-05-13 16:53:20+00) -- the timestamp is not
  a separate substring, it's the top 32 bits of a wider decoded value. New p_tt_discard_bits
  (right-shift the decoded field by this many bits before treating it as the time count;
  0 for cuid/ULID, 128 for KSUID) and p_tt_epoch (the zero-point; standard Unix epoch by
  default) parameters.

Both new formats, and the epoch/radix values, are verified against each format's own
source (segmentio/ksuid, ulid/spec, MongoDB's BSON reference), not guessed from names --
that mattered: cuid's timestamp field turned out not to be zero-padded (caught in the
previous commit), and this time a real arithmetic bug surfaced empirically while proving
the KSUID case at its actual 48-decimal-digit scale: _radix_encode's digit extraction used
floor(v_n / p_radix), and PostgreSQL's general numeric division computes non-terminating
quotients to a BOUNDED number of decimal digits -- fine at cuid's ~12-digit scale, wrong
(sometimes producing negative "digits") at KSUID's ~48-digit scale. Fixed with exact
integer div()/mod(), not floor(a/b) -- diagnosed by hand-tracing the encode loop against
an independently-computed Horner sum until the exact step it diverged at.

_radix_decode/_radix_encode also widen from bigint to numeric, since KSUID's whole-payload
value overflows a 64-bit bigint by nearly 100 decimal digits.

3 new pgpm.config columns (text_time_alphabet/discard_bits/epoch), threaded through
_encode/_decode and every call site, _transmute's validation (alphabet length must match
radix, must have no repeated characters), and check_text_time's plausibility sampling
(the shape-check character class now builds from a custom alphabet when supplied).

The existing frontier_data_only mutation and bench/frontier_drought.sh needed their
find/replace patterns and generated code re-synced to the now-7-trailing-param call
sites; ./test.sh discriminate confirms all 8 guards, including this one for both uuidv7
and text_time, still fail correctly against their reverted defects.

docs/guide.md's "Pick the kind" section now has all four ready-to-use recipes (cuid v1,
ULID, KSUID, MongoDB ObjectId) -- concrete parameter values, not an instruction to go
derive them from a generator's source.

+++ tests/86_text_time_codec_test.sql
@@ -12,8 +12,8 @@ select plan(10);
 
 -- ---- _radix_decode / _radix_encode: the bottom primitive, base-N <-> bigint, lowercase 0-9a-z ----
 
-select is(pgpm._radix_decode('z', 36), 35::bigint, 'radix_decode: single digit z (base36) = 35');
-select is(pgpm._radix_decode('10', 36), 36::bigint, 'radix_decode: 10 (base36) = 36 (Horner, not concatenation)');
+select is(pgpm._radix_decode('z', 36), 35::numeric, 'radix_decode: single digit z (base36) = 35');
+select is(pgpm._radix_decode('10', 36), 36::numeric, 'radix_decode: 10 (base36) = 36 (Horner, not concatenation)');
 select is(pgpm._radix_encode(35, 36, 1), 'z', 'radix_encode: 35 (base36), width 1 = z');
 select is(pgpm._radix_encode(36, 36, 2), '10', 'radix_encode: 36 (base36), width 2 = 10, zero-padded');
